Added confirmation message

+++ src/app/mylists/mylists.component.ts
@@ -24,8 +24,10 @@ export class MylistsComponent implements OnInit {
     }
 
     delete(id: number): void {
-        this.listService.deleteList(id).subscribe((data) => {
-            location.reload();
-        });
+        if (confirm('Are you sure you want to delete this custom list ?')) {
+            this.listService.deleteList(id).subscribe((data) => {
+                location.reload();
+            });
+        }
     }
 }
